requirements.txt
numpy
pandas
torch
torchvision
Pillow
scikit-learn
seaborn
matplotlib
joblib
